get active backend using write_url
